+++ src/main/java/radon/jujutsu_kaisen/ability/disaster_plants/WoodShield.java
@@ -45,7 +45,7 @@ public class WoodShield extends Summon<WoodShieldEntity> {
     }
 
     @Override
-    protected WoodShieldEntity summon(int index, LivingEntity owner) {
+    protected WoodShieldEntity summon(LivingEntity owner) {
         return new WoodShieldEntity(owner);
     }